Command/speaker_recognition/vgg_speaker_recognition/requirements.txt
absl-py==0.7.1
aniso8601==8.1.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.8.0
audioread==2.1.8
bleach==1.5.0
boto3==1.10.47
botocore==1.13.50
certifi==2020.12.5
cffi==1.13.2
chardet==4.0.0
click==7.1.2
cryptography==2.1.4
decorator==4.4.1
docker==4.4.1
docutils==0.15.2
english==2020.7.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
flask-url-map-serializer==0.0.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.35.0
grpcio-tools==1.35.0
h5py==2.9.0
html5lib==0.9999999
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.9.4
joblib==0.14.1
Keras==2.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
librosa==0.7.0
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
monotonic==1.5
mpu==0.20.0
neobolt==1.7.17
neotime==1.7.4
numba==0.46.0
numpy==1.18.2
packaging==20.9
pandas==1.0.3
pansi==2020.7.3
passlib==1.7.4
pika==1.1.0
prompt-toolkit==2.0.10
protobuf==3.8.0
psycopg2-binary==2.8.6
py2neo==2020.0.0
pycparser==2.19
pycrypto==2.6.1
Pygments==2.7.4
pymilvus==0.2.14
pyparsing==2.4.7
python-dateutil==2.8.0
python-dotenv==0.15.0
python-magic==0.4.18
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
requests==2.25.1
resampy==0.2.2
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.3.3
SecretStorage==2.3.1
six==1.15.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.40.2
ujson==4.0.2
urllib3==1.24.3
watchdog==1.0.2
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
